requirements.txt
altgraph==0.17.5
customtkinter==5.2.2
darkdetect==0.8.0
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
setuptools==82.0.1
